refactor: deprecate ibbtc png files in favor of sgv versions
