+++ src/components/Sidebar/index.tsx
@@ -17,7 +17,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import { ExpandMore } from '@material-ui/icons';
 import { SITE_VERSION, NETWORK_LIST, FLAGS } from 'config/constants';
 import NetworkWidget from 'components-v2/common/NetworkWidget';
-import ibBTCLogo from 'assets/tokens/ibBTC.png';
+import ibBTCLogo from 'assets/sidebar/ibBTC.svg';
 import { Route } from 'mobx-router';
 import { RootStore } from 'mobx/store';
